arreglos del user manager

+++ public/js/userManager.js
@@ -1,4 +1,5 @@
 const btnRole = document.querySelectorAll(".btn-changeRole")
+//const newRole = document.querySelectorAll(".roles-options").value
 const btnDelete = document.querySelectorAll(".btn-delete")
 const btnDeleteInactiveUsers = document.getElementById("btn-deleteInactiveUsers")
 
@@ -8,8 +9,9 @@ const usersManager = () => {
         btn.addEventListener("click", async (e) => {
             e.preventDefault()
             const ul = e.target.closest('ul')
+            const li = e.target.closest("li")
+            const newRole = li.children[0].value
             const email = ul.dataset.email
-            const newRole = document.getElementById("roles").value
 
             const result = await fetch(`/api/users/premium/${email}/${newRole}`)
 
@@ -31,7 +33,18 @@ const usersManager = () => {
                     location.reload();
                 }, 2000)
             } else {
-                console.log("not");
+                Toastify({
+                    text: `Error al cambiar el rol`,
+                    duration: 3000,
+                    className: "info",
+                    close: true,
+                    gravity: "top",
+                    position: "right",
+                    stopOnFocus: true,
+                    style: {
+                        background: "linear-gradient(to left, #b00017, #5e1f21)"
+                    }
+                }).showToast();
             }
 
         })
@@ -55,7 +68,6 @@ const usersManager = () => {
             }).then(async (result) => {
                 if (result.isConfirmed) {
 
-
                     const result = await fetch(`/api/users/deleteUser/${email}`, {
                         method: 'DELETE',
                         headers: {
@@ -73,8 +85,20 @@ const usersManager = () => {
                         setTimeout(() => {
                             location.reload();
                         }, 2000)
+
                     } else {
-                        console.log(not);
+                        Toastify({
+                            text: `Error al eliminar el usuario`,
+                            duration: 3000,
+                            className: "info",
+                            close: true,
+                            gravity: "top",
+                            position: "right",
+                            stopOnFocus: true,
+                            style: {
+                                background: "linear-gradient(to left, #b00017, #5e1f21)"
+                            }
+                        }).showToast();
                     }
 
                 }
@@ -86,18 +110,53 @@ const usersManager = () => {
     btnDeleteInactiveUsers.addEventListener("click", async (e) => {
         e.preventDefault()
 
-        const result = await fetch(`/api/users`, {
-            method: 'DELETE',
-            headers: {
-                'Content-Type': 'application/json',
+        Swal.fire({
+            title: 'Seguro quieres eliminar todos los usuarios inactivos?',
+            text: "No podras revertir el cambio",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Si, eliminar!'
+        }).then(async (result) => {
+            if (result.isConfirmed) {
+
+                const result = await fetch(`/api/users`, {
+                    method: 'DELETE',
+                    headers: {
+                        'Content-Type': 'application/json',
+                    }
+                })
+
+                if (result.ok) {
+                    Swal.fire(
+                        'Eliminado!',
+                        'Los usuarios han sido eliminado.',
+                        'success'
+                    )
+
+                    setTimeout(() => {
+                        location.reload();
+                    }, 2000)
+
+                } else {
+                    Toastify({
+                        text: `Error al eliminar los usuarios`,
+                        duration: 3000,
+                        className: "info",
+                        close: true,
+                        gravity: "top",
+                        position: "right",
+                        stopOnFocus: true,
+                        style: {
+                            background: "linear-gradient(to left, #b00017, #5e1f21)"
+                        }
+                    }).showToast();
+                }
+
             }
         })
 
-        if (result.ok) {
-            console.log("si");
-        } else {
-            console.log("no");
-        }
     })
 }
 usersManager()
